+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## Version 1.0.0
 - Initial version
 - Download music from YouTube Music to user's computer
+
+## Version 1.1.0
+- Added in-app notification text about downloading status
+- Added function to automatically check link before downloading
+- Added in-app logging
+- App window now cannot be resized for UI consistency
+- UI fixes
+- Code fixing & cleaning